feat(commands/): Added commands to generate dummy users and do a simple import of all the domain xlsx files
